move useGate to separate module and tsify
